pyproject.toml
[project]
name = "rbac-gitops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[tool.pyright]
typeCheckingMode = "basic"

